requirements.txt
Django==5.1
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
openai==1.12.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
pydantic==2.9.2
dj-database-url==2.1.0
